chore(scripts): extract shared frontmatter parser into _lib
